+++ Documentation/WhatsNew.md
@@ -36,7 +36,7 @@ Milestone 3
   * New ezVariant type that is much more powerful and useful than the previous implementation.
   * Added ezConversionUtils class for converting strings to int/float/bool.
   * Added optional *.natvis support for Visual Studio 2012. This enables much more comfortable inspection of basic types in the debugger.
-  
+  * The logging system has been improved. You can now log locally into another system (useful for tools) or you can replace the default implementation entirely.
 
 Milestone 2
 -----------